+++ app/src/main/java/com/coai/samin_total/Oxygen/SetOxygenViewData.kt
@@ -7,8 +7,8 @@ data class SetOxygenViewData(
     var sensorType: String = "LOX-02",
     var usable: Boolean = true,
     var isAlert: Boolean = false,
-    var setValue: Int = 0,
-    var setMinValue: Int = 20,
+    var setValue: Float = 0f,
+    var setMinValue: Float = 20f,
     var setMaxValue: Int = 23,
     val modelByte: Byte = 4,
     var heartbeatCount : UByte = 0u
